Refactor, added error break for no tiers
