+++ src/main.rs
@@ -1,19 +1,17 @@
 mod config;
-use chrono::{offset::Utc, DateTime};
 use colored::Colorize;
-use serde::{Deserialize, Serialize};
-use std::str::FromStr;
 use structopt::StructOpt;
-use uuid_5::Uuid;
 
 extern crate rand;
 use rand::thread_rng;
 use rand::Rng;
 
 use fake::faker::company::en::*;
-use fake::{Dummy, Fake, Faker};
+use fake::{Fake, Faker};
 
 use std::{thread, time};
+
+mod dsm;
 #[derive(Debug, StructOpt)]
 #[structopt()]
 pub struct Opt {
@@ -23,63 +21,6 @@ pub struct Opt {
     /// Number data source that will be injected. Number needs to be bigger than 1.
     #[structopt(short = "l", long, name = "limit")]
     dsm_limit: Option<i32>,
-}
-
-#[derive(Serialize, Debug)]
-pub struct DataSourceEntity {
-    tier_id: Uuid,
-    parent_id: Option<Uuid>,
-    name: String,
-    note: Option<String>,
-    status: EntityStatus,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize, Dummy)]
-#[serde(rename_all = "lowercase")]
-pub enum EntityStatus {
-    Published,
-    Disabled,
-}
-
-impl FromStr for EntityStatus {
-    type Err = &'static str;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "published" => Ok(EntityStatus::Published),
-            "disabled" => Ok(EntityStatus::Disabled),
-            _ => Err("Invalid status value"),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct TierRecord {
-    app_slug: String,
-    parent_id: Option<Uuid>,
-    id: Uuid,
-    name: String,
-    note: Option<String>,
-    status: TierStatus,
-    created_at: DateTime<Utc>,
-    updated_at: Option<DateTime<Utc>>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum TierStatus {
-    Published,
-    Disabled,
-}
-
-impl TierStatus {
-    pub fn from(text: &String) -> Self {
-        match text.as_str() {
-            "published" => Self::Published,
-            // currently allowing an invalid default to Disabled
-            _ => Self::Disabled,
-        }
-    }
 }
 
 async fn process(opt: &Opt, token: String) {
@@ -102,20 +43,23 @@ async fn process(opt: &Opt, token: String) {
     for app in 0..opt.app_slugs.len() {
         println!("APP: {:?}", &opt.app_slugs[app]);
         // Get number of tiers
-        let fetched_tiers = tiers(&opt.app_slugs[app], token.clone()).await;
-
+        let fetched_tiers = dsm::tier::tiers(&opt.app_slugs[app], token.clone()).await;
+        if fetched_tiers.tiers.len() <= 0 {
+            println!("{}", "ERROR: There isn't any tiers for this app.".red());
+            break;
+        }
         // Create Entities for top tier in case there isn't any.
         let mut rand_num_asset: i32 = rng.gen_range(1..limit);
         for _asset in 0..rand_num_asset {
-            let fake_dse = DataSourceEntity {
+            let fake_dse = dsm::entity::DataSourceEntity {
                 tier_id: fetched_tiers.tiers[0].id,
                 parent_id: None,
                 name: Buzzword().fake(),
                 note: CatchPhase().fake(),
-                status: Faker.fake::<EntityStatus>(),
+                status: Faker.fake::<dsm::entity::EntityStatus>(),
             };
             // post entity
-            post_entity(&opt.app_slugs[app], vec![fake_dse], token.clone()).await;
+            dsm::entity::post_entity(&opt.app_slugs[app], vec![fake_dse], token.clone()).await;
         }
 
         println!(
@@ -128,7 +72,7 @@ async fn process(opt: &Opt, token: String) {
         // Create Entities for each tiers after top tier.
         for tier in 1..fetched_tiers.tiers.len() {
             // Get entities in the tier before to set up as parents
-            let entities = get_entities(
+            let entities = dsm::entity::get_entities(
                 &opt.app_slugs[app],
                 token.clone(),
                 fetched_tiers.tiers[tier - 1].id,
@@ -139,101 +83,28 @@ async fn process(opt: &Opt, token: String) {
             for entity in entities.assets {
                 // Creating random number of entities for tier
                 rand_num_asset = rng.gen_range(1..limit);
-
+                println!(
+                    "CREATING {:?} entities for tier {:?}...",
+                    rand_num_asset, tier
+                );
                 for _rand_asset in 0..rand_num_asset {
-                    let fake_dse = DataSourceEntity {
+                    let fake_dse = dsm::entity::DataSourceEntity {
                         tier_id: fetched_tiers.tiers[tier].id,
                         parent_id: Some(entity.id),
                         name: Buzzword().fake(),
                         note: CatchPhase().fake(),
-                        status: Faker.fake::<EntityStatus>(),
+                        status: Faker.fake::<dsm::entity::EntityStatus>(),
                     };
                     // post entity
-                    post_entity(&opt.app_slugs[app], vec![fake_dse], token.clone()).await;
+                    dsm::entity::post_entity(&opt.app_slugs[app], vec![fake_dse], token.clone())
+                        .await;
                 }
             }
             // Need delay between each entity creation in a tier
-            println!(
-                "\nCREATING {:?} entities for tier {:?}...(30 secs)",
-                rand_num_asset, tier
-            );
+            println!("\nPOST delay...(30 secs)");
             let post_delay = time::Duration::from_millis(30000);
             thread::sleep(post_delay);
         }
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-pub struct TiersResult {
-    #[serde(alias = "records")]
-    pub tiers: Vec<TierRecord>,
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-pub struct EntitiesResult {
-    #[serde(alias = "records")]
-    pub assets: Vec<EntityRecord>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct EntityRecord {
-    app_slug: String,
-    tier_id: Uuid,
-    parent_id: Option<Uuid>,
-    id: Uuid,
-    name: String,
-    note: Option<String>,
-    status: EntityStatus,
-    created_at: DateTime<Utc>,
-    updated_at: Option<DateTime<Utc>>,
-}
-
-pub async fn tiers(app_slug: &str, token: String) -> TiersResult {
-    let hostname = config::api::dsm::connection_url();
-    let url = format!("{}/v1/{}/tiers", hostname, app_slug);
-    let client = reqwest::Client::new();
-    let response = client.get(&url).bearer_auth(token).send().await.unwrap();
-
-    let json_response;
-    if response.status().is_success() {
-        json_response = response.json::<TiersResult>().await.unwrap();
-    } else {
-        json_response = TiersResult { tiers: Vec::new() };
-    }
-    json_response
-}
-
-pub async fn get_entities(app_slug: &str, token: String, tier_id: Uuid) -> EntitiesResult {
-    let hostname = config::api::dsm::connection_url();
-    let url = format!("{}/v1/{}/assets?tier_id={}", hostname, app_slug, tier_id);
-    let client = reqwest::Client::new();
-    let response = client.get(&url).bearer_auth(token).send().await.unwrap();
-
-    let json_response;
-    if response.status().is_success() {
-        json_response = response.json::<EntitiesResult>().await.unwrap();
-    } else {
-        json_response = EntitiesResult { assets: Vec::new() };
-    }
-    json_response
-}
-
-pub async fn post_entity(app_slug: &str, fake_dse: Vec<DataSourceEntity>, token: String) {
-    let hostname = config::api::dsm::connection_url();
-    let url = format!("{}/v1/{}/assets", hostname, app_slug);
-
-    let client = reqwest::Client::new();
-
-    let response = client
-        .post(&url)
-        .bearer_auth(token)
-        .json(&fake_dse)
-        .send()
-        .await
-        .unwrap();
-
-    if !response.status().is_success() {
-        println!("{:?}", "ERROR: Issue with post entity".red());
     }
 }
